Move detectors from detekt package to domain packages
complexity, exceptions, naming, style
