+++ slack-lint-checks/src/main/java/slack/lint/SlackIssueRegistry.kt
@@ -7,31 +7,23 @@ import com.android.tools.lint.client.api.Vendor
 import com.android.tools.lint.detector.api.CURRENT_API
 import com.android.tools.lint.detector.api.Issue
 import com.google.auto.service.AutoService
+import slack.lint.complexity.CyclomaticComplexMethodDetector
+import slack.lint.complexity.LongMethodDetector
+import slack.lint.complexity.LongParameterListDetector
+import slack.lint.complexity.NestedBlockDepthDetector
+import slack.lint.complexity.TooManyFunctionsDetector
 import slack.lint.denylistedapis.DenyListedApiDetector
-import slack.lint.detekt.CyclomaticComplexMethodDetector
-import slack.lint.detekt.FunctionNamingDetector
-import slack.lint.detekt.FunctionOnlyReturningConstantDetector
-import slack.lint.detekt.InvalidPackageDeclarationDetector
-import slack.lint.detekt.LongMethodDetector
-import slack.lint.detekt.LongParameterListDetector
-import slack.lint.detekt.MagicNumberDetector
-import slack.lint.detekt.NestedBlockDepthDetector
-import slack.lint.detekt.ReturnCountDetector
-import slack.lint.detekt.SwallowedExceptionDetector
-import slack.lint.detekt.ThrowingExceptionsWithoutMessageDetector
-import slack.lint.detekt.TooGenericExceptionCaughtDetector
-import slack.lint.detekt.TooGenericExceptionThrownDetector
-import slack.lint.detekt.TooManyFunctionsDetector
-import slack.lint.detekt.TopLevelPropertyNamingDetector
-import slack.lint.detekt.UnderscoresInNumericLiteralsDetector
-import slack.lint.detekt.UnnecessaryAbstractClassDetector
-import slack.lint.detekt.UnnecessaryBackticksDetector
-import slack.lint.detekt.UnusedPrivateMemberDetector
-import slack.lint.detekt.VarCouldBeValDetector
 import slack.lint.eithernet.DoNotExposeEitherNetInRepositoriesDetector
+import slack.lint.exceptions.SwallowedExceptionDetector
+import slack.lint.exceptions.ThrowingExceptionsWithoutMessageDetector
+import slack.lint.exceptions.TooGenericExceptionCaughtDetector
+import slack.lint.exceptions.TooGenericExceptionThrownDetector
 import slack.lint.inclusive.InclusiveNamingChecker
 import slack.lint.mocking.ErrorProneDoNotMockDetector
 import slack.lint.mocking.MockDetector
+import slack.lint.naming.FunctionNamingDetector
+import slack.lint.naming.InvalidPackageDeclarationDetector
+import slack.lint.naming.TopLevelPropertyNamingDetector
 import slack.lint.parcel.ParcelizeFunctionPropertyDetector
 import slack.lint.resources.FullyQualifiedResourceDetector
 import slack.lint.resources.MissingResourceImportAliasDetector
@@ -39,6 +31,14 @@ import slack.lint.resources.WrongResourceImportAliasDetector
 import slack.lint.retrofit.RetrofitUsageDetector
 import slack.lint.rx.RxObservableEmitDetector
 import slack.lint.rx.RxSubscribeOnMainDetector
+import slack.lint.style.FunctionOnlyReturningConstantDetector
+import slack.lint.style.MagicNumberDetector
+import slack.lint.style.ReturnCountDetector
+import slack.lint.style.UnderscoresInNumericLiteralsDetector
+import slack.lint.style.UnnecessaryAbstractClassDetector
+import slack.lint.style.UnnecessaryBackticksDetector
+import slack.lint.style.UnusedPrivateMemberDetector
+import slack.lint.style.VarCouldBeValDetector
 import slack.lint.text.SpanMarkPointMissingMaskDetector
 import slack.lint.ui.DoNotCallViewToString
 import slack.lint.ui.ItemDecorationViewBindingDetector
