+++ slack-lint-checks/src/main/java/slack/lint/complexity/LongMethodDetector.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2025 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package slack.lint.detekt
+package slack.lint.complexity
 
 import com.android.tools.lint.client.api.UElementHandler
 import com.android.tools.lint.detector.api.Category
@@ -10,58 +10,56 @@ import com.android.tools.lint.detector.api.JavaContext
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import org.jetbrains.uast.UElement
+import org.jetbrains.uast.UExpression
 import org.jetbrains.uast.UMethod
-import org.jetbrains.uast.UReturnExpression
-import org.jetbrains.uast.visitor.AbstractUastVisitor
 import slack.lint.util.IntLintOption
 import slack.lint.util.OptionLoadingDetector
 import slack.lint.util.sourceImplementation
 
-class ReturnCountDetector(private val maxOption: IntLintOption = IntLintOption(MAX_RETURNS)) :
-  OptionLoadingDetector(maxOption), SourceCodeScanner {
+class LongMethodDetector(private val thresholdOption: IntLintOption = IntLintOption(THRESHOLD)) :
+  OptionLoadingDetector(thresholdOption), SourceCodeScanner {
 
   override fun getApplicableUastTypes(): List<Class<out UElement>> = listOf(UMethod::class.java)
 
   override fun createUastHandler(context: JavaContext): UElementHandler {
     return object : UElementHandler() {
       override fun visitMethod(node: UMethod) {
-        var returnCount = 0
-        node.accept(
-          object : AbstractUastVisitor() {
-            override fun visitReturnExpression(node: UReturnExpression): Boolean {
-              returnCount++
-              return super.visitReturnExpression(node)
-            }
-          }
-        )
-        if (returnCount > maxOption.value) {
+        val body = node.uastBody ?: return
+        val lineCount = countLines(context, body)
+        if (lineCount > thresholdOption.value) {
           context.report(
             ISSUE,
             node,
             context.getNameLocation(node),
-            "Function has $returnCount return statements, exceeding the limit of ${maxOption.value}.",
+            "Function body has $lineCount lines, exceeding the limit of ${thresholdOption.value}.",
           )
         }
       }
     }
   }
 
+  private fun countLines(context: JavaContext, body: UExpression): Int {
+    val sourcePsi = body.sourcePsi ?: return 0
+    val text = sourcePsi.text
+    return text.count { it == '\n' }
+  }
+
   companion object {
-    internal val MAX_RETURNS =
-      IntOption("max-returns", "Maximum number of return statements allowed per function.", 4)
+    private val THRESHOLD =
+      IntOption("threshold", "Maximum number of lines allowed in a function body.", 120)
 
     val ISSUE =
       Issue.create(
-          id = "ReturnCount",
-          briefDescription = "Function has too many return statements",
+          id = "LongMethod",
+          briefDescription = "Function body is too long",
           explanation =
-            "Functions with many return statements are harder to follow. " +
-              "Consider refactoring to reduce the number of exit points.",
+            "Long functions are harder to understand, test, and maintain. " +
+              "Consider extracting parts of the function into smaller, well-named functions.",
           category = Category.CORRECTNESS,
           priority = 5,
           severity = Severity.WARNING,
-          implementation = sourceImplementation<ReturnCountDetector>(),
+          implementation = sourceImplementation<LongMethodDetector>(),
         )
-        .setOptions(listOf(MAX_RETURNS))
+        .setOptions(listOf(THRESHOLD))
   }
 }
