+++ slack-lint-checks/src/main/java/slack/lint/naming/FunctionNamingDetector.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2025 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package slack.lint.detekt
+package slack.lint.naming
 
 import com.android.tools.lint.client.api.UElementHandler
 import com.android.tools.lint.detector.api.Category
@@ -9,40 +9,39 @@ import com.android.tools.lint.detector.api.JavaContext
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import com.android.tools.lint.detector.api.StringOption
-import org.jetbrains.uast.UClass
+import com.android.tools.lint.detector.api.isKotlin
 import org.jetbrains.uast.UElement
+import org.jetbrains.uast.UMethod
 import slack.lint.util.OptionLoadingDetector
 import slack.lint.util.StringSetLintOption
 import slack.lint.util.hasAnyAnnotation
 import slack.lint.util.sourceImplementation
 
-class UnnecessaryAbstractClassDetector(
+private val FUNCTION_NAME_PATTERN = Regex("^[a-z][a-zA-Z0-9]*$")
+
+class FunctionNamingDetector(
   private val ignoreAnnotatedOption: StringSetLintOption = StringSetLintOption(IGNORE_ANNOTATED)
 ) : OptionLoadingDetector(ignoreAnnotatedOption), SourceCodeScanner {
 
-  override fun getApplicableUastTypes(): List<Class<out UElement>> = listOf(UClass::class.java)
+  override fun getApplicableUastTypes(): List<Class<out UElement>> = listOf(UMethod::class.java)
 
   override fun createUastHandler(context: JavaContext): UElementHandler {
+    if (!isKotlin(context.psiFile)) return UElementHandler.NONE
     return object : UElementHandler() {
-      override fun visitClass(node: UClass) {
-        if (!node.isAbstract()) return
+      override fun visitMethod(node: UMethod) {
+        if (node.isConstructor) return
+        val name = node.name
+        if (name.startsWith("`")) return
+        if (FUNCTION_NAME_PATTERN.matches(name)) return
         if (node.hasAnyAnnotation(ignoreAnnotatedOption.value)) return
-
-        val hasAbstractMembers = node.methods.any { method -> context.evaluator.isAbstract(method) }
-        if (!hasAbstractMembers) {
-          context.report(
-            ISSUE,
-            node,
-            context.getNameLocation(node),
-            "Abstract class `${node.name}` has no abstract members. Remove the `abstract` modifier or add abstract members.",
-          )
-        }
+        context.report(
+          ISSUE,
+          node,
+          context.getNameLocation(node),
+          "Function name `$name` should start with a lowercase letter and use camelCase.",
+        )
       }
     }
-  }
-
-  private fun UClass.isAbstract(): Boolean {
-    return javaPsi.modifierList?.hasModifierProperty("abstract") == true
   }
 
   companion object {
@@ -50,21 +49,21 @@ class UnnecessaryAbstractClassDetector(
       StringOption(
         "ignore-annotated",
         "Comma-separated list of annotation simple names to ignore.",
-        "Module",
-        "Abstract classes annotated with these are excluded (e.g. Dagger modules).",
+        "Composable,Test",
+        "Functions annotated with these annotations are excluded from the naming check.",
       )
 
     val ISSUE =
       Issue.create(
-          id = "UnnecessaryAbstractClass",
-          briefDescription = "Abstract class has no abstract members",
+          id = "FunctionNaming",
+          briefDescription = "Function name does not follow naming conventions",
           explanation =
-            "An abstract class without abstract members doesn't need to be abstract. " +
-              "Consider making it a concrete class or an interface, or add abstract members.",
+            "Function names should start with a lowercase letter and use camelCase. " +
+              "Exceptions can be configured via the ignore-annotated option.",
           category = Category.CORRECTNESS,
           priority = 5,
           severity = Severity.WARNING,
-          implementation = sourceImplementation<UnnecessaryAbstractClassDetector>(),
+          implementation = sourceImplementation<FunctionNamingDetector>(),
         )
         .setOptions(listOf(IGNORE_ANNOTATED))
   }
