+++ slack-lint-checks/src/main/java/slack/lint/naming/InvalidPackageDeclarationDetector.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2025 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package slack.lint.detekt
+package slack.lint.naming
 
 import com.android.tools.lint.detector.api.Category
 import com.android.tools.lint.detector.api.Context
